refactor: [FE] QuizResult 컴포넌트 분리
